[Test] IR Deserialization: Hide dispatch receiver type
This is necessary to avoid discrepancies in IR deserialization tests
between dumps of IR produced by Fir2Ir and by the deserializer.
Because dispatch receivers for the corresponding fake overrides
in those two groups of IR sometimes are not identical (though, always
compatible). And, as far as this has no negative effect on the compiler,
the algorithm(s) of fake override generation are not going to be
fixed.

^KT-73171