[PowerAssert] Add IR origin for implicit call arguments
The compiler frontend tracks which receiver - if any - is specified
explicitly. This information is not maintained when converted to IR.
This means a plugin like power-assert is unable to ignore arguments
which are supplied implicitly.

Add a new IrStatementOrigin which indicates when an argument is
implicit. This can be used by power-assert to determine what arguments
can be ignored and to also fix alignment of some operator calls.

Unfortunately, IrGetObjectValue doesn't contain an origin, so these
elements cannot be marked as implicit. However, since object access
shouldn't be shown, these elements can always be ignored.

^KT-73870 Fixed
^KT-73897 Fixed
^KT-73898 Fixed